Move call context to its own file
